integrate views with router, remove dead assets, adjust toolbar links, remove #
