+++ web/src/router/index.js
@@ -22,6 +22,7 @@ const routes = [
 
 const router = new VueRouter({
   routes,
+  mode: 'history', // makes the # go away
 });
 
 export default router;
