+++ web/src/views/About.vue
@@ -1,5 +1,12 @@
 <template>
-  <div class="about">
-    <h1>This is an about page</h1>
+  <div>
+    <h1>About this project</h1>
+    <p>
+      Lorem ipsum dolor sit amet consectetur, adipisicing elit.
+      Earum eum, quos quibusdam deleniti temporibus
+      harum. Expedita natus excepturi optio, sequi, maiores, earum
+      labore at totam a corporis quas itaque
+      rerum.
+    </p>
   </div>
 </template>
